Make test util file for fake rtsp cam
